update based on comments
