+++ core/src/main/kotlin/com/expediagroup/sdk/core/client/BaseXapClient.kt
@@ -23,7 +23,6 @@ import com.expediagroup.sdk.core.plugin.authentication.strategy.AuthenticationSt
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.engine.okhttp.OkHttp
-import okhttp3.OkHttpClient
 
 /**
  * The integration point between the SDK Core and the product SDKs.
@@ -42,7 +41,7 @@ abstract class BaseXapClient(
             XapConfigurationProvider
         )
 
-    private val engine: HttpClientEngine = clientConfiguration.okHttpClient?.let {
+    private val engine: HttpClientEngine = _configurationProvider.okHttpClient?.let {
         OkHttp.create {
             preconfigured = it
         }
@@ -66,17 +65,5 @@ abstract class BaseXapClient(
 
     /** A [BaseXapClient] builder with ability to pass a custom okhttp client. */
     @Suppress("unused", "UnnecessaryAbstractClass") // This is used by the generated SDK clients.
-    abstract class BuilderWithHttpClient<SELF : Client.Builder<SELF>> : Client.Builder<SELF>() {
-        protected var okHttpClient: OkHttpClient? = null
-
-        @Suppress("UNCHECKED_CAST")
-        override fun self(): SELF = this as SELF
-
-        /** Sets the [OkHttpClient] to use for the [BaseXapClient]. */
-        fun okHttpClient(okHttpClient: OkHttpClient): SELF {
-            this.okHttpClient = okHttpClient
-            return self()
-        }
-
-    }
+    abstract class BuilderWithHttpClient<SELF : Client.BuilderWithHttpClient<SELF>> : Client.BuilderWithHttpClient<SELF>()
 }
