+++ core/src/main/kotlin/com/expediagroup/sdk/core/configuration/ClientConfiguration.kt
@@ -16,6 +16,7 @@
 package com.expediagroup.sdk.core.configuration
 
 import com.expediagroup.sdk.core.configuration.provider.RuntimeConfigurationProvider
+import okhttp3.OkHttpClient
 
 interface ClientConfiguration {
     val key: String?
@@ -26,6 +27,7 @@ interface ClientConfiguration {
     val socketTimeout: Long?
     val maskedLoggingHeaders: Set<String>?
     val maskedLoggingBodyFields: Set<String>?
+    val okHttpClient: OkHttpClient?
 
     /** Build a [RuntimeConfigurationProvider] from a [ClientConfiguration]. */
     fun toProvider(): RuntimeConfigurationProvider =
@@ -37,6 +39,7 @@ interface ClientConfiguration {
             connectionTimeout = connectionTimeout,
             socketTimeout = socketTimeout,
             maskedLoggingHeaders = maskedLoggingHeaders,
-            maskedLoggingBodyFields = maskedLoggingBodyFields
+            maskedLoggingBodyFields = maskedLoggingBodyFields,
+            okHttpClient = okHttpClient
         )
 }
