+++ core/src/main/kotlin/com/expediagroup/sdk/core/configuration/ExpediaGroupClientConfiguration.kt
@@ -17,6 +17,7 @@ package com.expediagroup.sdk.core.configuration
 
 import com.expediagroup.sdk.core.client.ExpediaGroupClient
 import com.expediagroup.sdk.core.configuration.provider.RuntimeConfigurationProvider
+import okhttp3.OkHttpClient
 
 /**
  * Configuration for the [ExpediaGroupClient].
@@ -30,6 +31,7 @@ import com.expediagroup.sdk.core.configuration.provider.RuntimeConfigurationProv
  * @property maskedLoggingHeaders The headers to be masked in logging.
  * @property maskedLoggingBodyFields The body fields to be masked in logging.
  * @property authEndpoint The API endpoint to use for authentication.
+ * @property okHttpClient The okhttp client to be used by the sdk.
  */
 data class ExpediaGroupClientConfiguration(
     override val key: String? = null,
@@ -40,6 +42,7 @@ data class ExpediaGroupClientConfiguration(
     override val socketTimeout: Long? = null,
     override val maskedLoggingHeaders: Set<String>? = null,
     override val maskedLoggingBodyFields: Set<String>? = null,
+    override val okHttpClient: OkHttpClient? = null,
     val authEndpoint: String? = null
 ) : ClientConfiguration {
     /** Build a [RuntimeConfigurationProvider] from an [ExpediaGroupClientConfiguration]. */
