+++ core/src/main/kotlin/com/expediagroup/sdk/core/configuration/RapidClientConfiguration.kt
@@ -39,5 +39,5 @@ data class RapidClientConfiguration(
     override val socketTimeout: Long? = null,
     override val maskedLoggingHeaders: Set<String>? = null,
     override val maskedLoggingBodyFields: Set<String>? = null,
-    val okHttpClient: OkHttpClient? = null
+    override val okHttpClient: OkHttpClient? = null
 ) : ClientConfiguration
